requirements.txt
asgiref==3.8.1
coverage==7.8.2
Django==5.2.1
django-cors-headers==4.4.0
django-filter==25.1
django-model-utils==5.0.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.10
factory_boy==3.3.3
Faker==37.3.0
inflection==0.5.1
iniconfig==2.1.0
model-bakery==1.20.4
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pytest==8.3.5
pytest-cov==6.1.1
pytest-django==4.11.1
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
setuptools==80.9.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
